[react] Fix poorly positioned React error messages
Summary:
We were incorrectly forgetting to propagate the `use_op` in non-invariant flows with React config calculation. This lead to weird error messages that were positioned in places that were not helpful and also omitted helpful context to fix the error.

I added a test that has similar behavior to the example in the support group that lead to this investigation.

Reviewed By: samwgoldman

Differential Revision: D13627309

fbshipit-source-id: 87519208129507ea1c36add5521af39f2a2a88eb