[Any] include implicit vs explicitness of top-level any types in normalizer output
Summary:
Include info about explict or implicitness of anys in normalizer output. This should affect things like dump-types and type-at-pos.

Because we want the types produced here to look like real types, I followed the example we use for imported types and classes and include the extra text only when the types appear at the top level.

Reviewed By: panagosg7

Differential Revision: D13658990

fbshipit-source-id: 94be526d309a625f57a0a4a48ec92b0860f397ae