Prevent exception in signature_builder_generate when processing private fields
Summary:
This diff prevents an exception reported in https://github.com/facebook/flow/issues/7355 that'll cause Flow 0.9.0 to always except on private fields.

The fix is in place until we fully understand whether private properties are part of the generated signature or not.

This diff needs to be followed up in T39206072 after the release.

Reviewed By: panagosg7

Differential Revision: D13695528

fbshipit-source-id: 3c4ece37f8bc31da79c8c29ab365070270915756