Improve error message for lower bounds incompatible with shapes
Summary:
The logic added in D13201962 makes the wildcard `(_, ShapeT _)` case an error
always by changing the shape's `tout` type to `empty`. However, the reasons are
reversed, so when we see `(null: $Shape<{}>)` we say that the object type is
incompatible with null, when we should say that null is incompatible with the
shape.

Instead of using `ObjAssignT` in the wildcard case, we can just error early. We
can also reuse the `EFunctionIncompatibleWithShape` error, which does not use
wording actually specific to functions.

Reviewed By: dsainati1

Differential Revision: D13708538

fbshipit-source-id: f44b193e36863deec7922001e3d4f28c1ecd47b7