Expose location type parameter on `error` type
Summary:
I realized that I need to pass around errors with concrete locations, instead of always concretizing right when we are rendering them. This is because we will not have the necessary state to concretize locations in the client, so we have to concretize on the server before sending errors to the client.

Because of this, we have to make it possible for other modules to refer to errors which contain concrete locs.

Reviewed By: mroch

Differential Revision: D13716645

fbshipit-source-id: e71359c344618acb234e1156e68d0762497dd725