Add .timeout() to `./tool test` TestSteps
Summary: In order to test parallelizable persistent commands, I needed to assert that sometimes the Flow server would NOT respond to a request within N seconds. However, step was using `ideRequestAndWaitUntilResponse`, which would wait indefinitely for a response. By adding a timeout, it made it easy for me to stop waiting after N seconds.

Reviewed By: panagosg7

Differential Revision: D13636165

fbshipit-source-id: d1767efcf3e9697c2212297c137c6ad131f9ccd6