Make Persistent_connection stateful
Summary:
Before this diff, `ServerEnv.env` carries around a list of persistent clients. This diff changes `ServerEnv.env` to carry around a list of persistent client IDs. Then, inside of `Persistent_connection`, we keep a ref with a map from client IDs to the currently active clients.

Why do this? In {D13617364}, it would be nice to be able to immediately respond to persistent commands like

* A command for which we've already received a cancellation
* No-op commands, like `DidSaveNotification`
* Unsupported commands (we just send back an error)

However when you handle something immediately you don't have the `env`. So even though we know the client_id, we couldn't actually send the response or log the right data.

I left a list of client IDs in `ServerEnv.env`. If we receive a new persistent connection in the middle of a recheck, we don't want to start streaming errors to that persistent connection until the next recheck.

Reviewed By: nmote

Differential Revision: D13617365

fbshipit-source-id: 1c0a7481c9931b6a5c3e8fdccb0f388edaff4c27