Allow persistent commands run in parallel to be canceled
Summary:
By forgetting to catch `Lwt.Canceled` in `handle_parallelizable_persistent`, a recheck would stop handling parallelizable commands as soon as one parallelizable command was canceled. We were properly catching it for ephemeral commands, so fixing that here.

Also, the catch is only needed for when we run a command in parallel. Commands run in serial shouldn't actually throw `Lwt.Canceled` (they handle it themselves).

Reviewed By: jbrown215

Differential Revision: D13738835

fbshipit-source-id: 01847f0f1cb3cffafb9b69103dc0627daa3bf94b