No longer cancel commands on file changes
Summary:
1. The monitor tells use when files have changed. Changed files cause us to cancel rechecks and commands
2. We might also be asked to force certain files to be checked as focused files or dependencies. Sometimes this happens due to commands like `flow force-recheck --focus` and sometimes this happens when we need some unchecked dependencies to be checked.

It's a little tricky to know which when to cancel rechecks and commands. During a recheck, we need to estimate what files will be checked by the end of the recheck. {D13696452} refines this a bit, but it's a little annoying to pass around the estimate.

And then there are situations like this:

1. We run the `didOpen` command
2. That command starts a recheck
3. During the recheck we start running an `autocomplete`.

What should happen if some file changes?

We can just get rid of the file-changes-cancel-commands behavior. Commands can still cancel themselves if they need more files checked, but they no longer need to watch for necessary rechecks. For the example above, if a file changes, now...

1. The `autocomplete` finishes normally
2. The recheck cancels itself and starts a new recheck
3. When the recheck finishes, the `didOpen` sends a response.

Reviewed By: jbrown215

Differential Revision: D13738837

fbshipit-source-id: 4066b5af35d9ad8f6fd66e246260745faa399e07