Better predict what files will be checked after a recheck
Summary:
During a recheck, if we discover even more files need to be rechecked, we cancel the recheck and start a larger recheck.

Earlier I noticed that sometimes we were canceling rechecks due to files which were already being rechecked. For example

1. We start a recheck
2. `autocomplete` needs `dependency.js` to be checked. So we cancel the recheck and start a new recheck including `dependency.js`
3. Another `autocomplete` also needs `dependency.js` to be checked. So again we cancel the recheck, even though this recheck would have already checked `dependency.js`.

{D13511798} fixed this issue by keeping track of which files were were forcing to be checked.

However, I noticed another issue.

1. We'd start up a lazy server using ide lazy mode
2. We'd open `foo.js`. We'd explicitly force `foo.js` to be focused, which would also implicitly check `dependency.js`
3. During the recheck, we'd send a `type-at-pos` in `foo.js`. This would notice that `dependency.js` hasn't been checked yet and would cause the recheck to be canceled.

The kernel of the problem was that we were underestimating which files would be checked by the end of the recheck. This diff tries to improve the estimate as the recheck computes what will actually be rechecked.

Reviewed By: jbrown215

Differential Revision: D13696452

fbshipit-source-id: b88bcb9980ce9c2d7ce4e3a36b36c689e763d76b