Add mutable fields to single_client
Summary:
The tests for {D13738834} version 2 were consistently failing. When testing `flow ide`, `textDocument/didOpen` was failing to return the errors in a file. That's because of code that looked like

```
    if Persistent_connection.client_did_open client ~files
    then did_open genv env client files
    else Lwt.return env  (* No new files were opened, so do nothing *)
```

`client_did_open` would modify `client` and save the new version to `Persistent_connection.active_clients`. However, `did_open` was being called with the old client.

Persistent clients are inherently stateful, since they are used to track state external to Flow. So I think this solution is cleaner, clearer, and less like to cause future bugs.

Reviewed By: jbrown215

Differential Revision: D13775265

fbshipit-source-id: 01cc22fc4e4e6976cc9bf5162e185fb4cd1e28b1