Immediately grab file contents
Summary:
Back when we ran all commands in order, we didn't need to worry about this. But now we have to worry about something like

1. didOpen
2. autocomplete
3. didChange

Since we process `didChange` immediately, when we run autocomplete we might have the wrong buffer.

To fix, let's just take a snapshot of the buffer as soon as the command comes in. Since these are stored as strings, this should only use more memory if the file actually changes.

Reviewed By: jbrown215

Differential Revision: D13738833

fbshipit-source-id: cd79e6b4d8d6a1de6e948da097bff8be641969fe