[hack] Remove unused ability to connect to shared mem as master
Summary:
This ability was added In D2796468, which was part of the work to support
Windows.

At the time, Hack's monitor process was responsible for initializing shared
memory, but the server process was the designated master process. To support
this, the `connect ~is_master:true` call scheme was added to allow for a kind
of ownership transfer.

Today, it is always the case that the master process initializes shared memory.
Since initialization does all the work of hh_connect already, all calls that
attempt to connect as master are superfluous: they clobber the globals with the
values that are already there and they re-mmap the shared heap.

It seems that confusion around this API has led to tests unnecessarily
connecting after init. To avoid this confusion, I am simply removing this API.

Reviewed By: gabelevi

Differential Revision: D13725783

fbshipit-source-id: 52e1e1d3a23c6e9944fe38c638d4866869616194