Convert errors to contain concrete locations upon collation
Summary:
This builds on the previous diff to push error location concretization even earlier. The motivation for this was to remove the unsafe location conversion in `error_suppressions.ml`, but it also seems a bit cleaner this way.

It's fine to concretize when we collate because we will re-collate after each recheck. In the meantime, we can store the errors with the concrete locs and use them to answer subsequent responses.

Reviewed By: jbrown215

Differential Revision: D13771120

fbshipit-source-id: 00e4903a70318f8b7959c3ff87902451a4ed6e87