Send errors with concrete locations down to the client
Summary:
Right now, all abstract locations have a concrete underlying representation. When that changes, additional context will be necessary in order to convert abstract locations to concrete locations. Only the server will have this context, so we have to convert locations before we send them down to the client.

This is also more palatable than the previous approach of converting locations ad-hoc as they are getting printed.

Reviewed By: jbrown215

Differential Revision: D13716644

fbshipit-source-id: 437594e2d80f87150592bfb3bdb598ca74df2224