[PR] Add TransitionEvent.
Summary:
Add `TransitionEvent` as a type and accept it as a parameter to `addEventListener` and other such functions.

I think this is correct? I might have missed something. I tried following the code for the other events such as the `ClipboardEvent`.

I also followed the interface on https://www.w3.org/TR/2017/WD-css-transitions-1-20171130/#interface-transitionevent to do this.
Pull Request resolved: https://github.com/facebook/flow/pull/6614

Reviewed By: dsainati1

Differential Revision: D13830433

Pulled By: nmote

fbshipit-source-id: 2fc66722572618b20bf30bce98feccc67e0cccc9