fix positioning of inexact-spread lint errors
Summary: `inexact-spread` lint errors were being placed at the site of the definition of the spreaded type rather than the place where the type was being used, breaking the invariant that the primary location of lints be in the file in which the error occurs. This also has the nice benefit of simplifying the error message for this lint.

Reviewed By: samwgoldman

Differential Revision: D13883265

fbshipit-source-id: 020c54cff00d9dbdab7a149215101a8a6c2e831e