Three-color coverage
Summary:
A node can now be one of three things w.r.t. coverage:

**Checked**: this corresponds to the existing "covered" mode. This includes the
following cases:
* All types whose top-level constructor is not `any`, `empty`, `|` (union) or
`&` (intersection).
* All types whose top-level constructor is `|` (union) and
(a) none of its members is **Any**, and
(b) at least one of its members is not **Empty**.
* All types whose top-level constructor is `&` (intersection) and
(a) none of its members is **Any**, and
(b) none of its members is **Empty**.

**Any**: this is one of the categories that were considered "uncovered" before. This
includes:
* The type `any`.
* All types whose top-level constructor is `|` (union) and any of its members is **Any**.
* All types whose top-level constructor is `&` (intersection) and any of its members is **Any**.

**Empty**: This corresponds to dead-code. Earlier this case was also considered
uncovered. Structurally this includes all remaining cases.

 ---

Like before type variables with multiple lower bounds are considered as the union of
those bounds.

Note that the definition is not symmetric between **Any** and **Empty** in aggregate types
(unions/intersections). The reason for this is to capture the idiomatic behavior of the
checker when `any` is part of those types. These types are now uncovered, due to the
relaxed checking of the following cases:
* Upper bounds of `T & any`:
```
declare var x: number & any;;
(x: string);  // no error
x = "";  // error
```
* Lower bounds of `T | any`:
```
declare var x: string | any;
if (0 < 1) { x = 1 }; // no error
(x: string);
```

This adds a blue color for the Empty case and red color for Any.

This adds a `"deadcode_count"` and a `"deadcode_locs"` section to the output,
that correspond to **Empty**.

he existing `"uncovered*"` sections are the sum of **Any** and **Empty** cases.

The above rules definition determines the composition operators between coverage kinds.

We still try to exit early when the coverage has been determined.

Reviewed By: dsainati1

Differential Revision: D13826267

fbshipit-source-id: fbccaa75fc99012867619d64059a959336370c75