Dedicated coverage module
Summary:
Moves coverage computation away from the normalizer and to a dedicated optimized visitor.
The goal is to have a fast enough way to compute coverage on merge and keep the result on
the server.

Some cases in which this differs from the previous behavior:

**Type definitions**

`coverage (TypeT t) = coverage t`

**Type Aliases**

We now descend in the body of a type alias. Before:
```
type A = any;
declare var a: A;
a;
```
`a` was considered covered.

Now it will be uncovered.

**`CharSetT`** is now covered

**Recursive types** where the recursion is not guarded by a constructor are uncovered,
since their semantics are not straight-forward.

E.g.
```
type T = T | number;
declare var x: T
if (0 < 1) { x = "" };
(x: number); // no error here
```

Reviewed By: dsainati1

Differential Revision: D13794124

fbshipit-source-id: 909c6fad234a58a4aebbe83ebdc5b21071143e05