[hack] Improve findRefs to capture class_meth without apostrophes
Summary:
Modify the `find_refs` task to be aware of function references that use apostrophes.

This occurs in a few types of cases:
* When using fun('global') to refer to a global function.
* When using class_meth(class, 'method') to refer to a static method within a class.
* When using inst_meth(object, 'method') to refer to an instance method within a specific instance of a class.

When using automated codemods, the `find_refs` task must capture the name of the method without the surrounding apostrophes.

To solve this:
* I added tests for each of the three use cases above.
* Added a new ocaml method on `Pos.t`, `shrink_by_one_char_both_sides`, which adjusts two `Pos.t` objects and removes apostrophes from capture
* Modified IdentifySymbolService to use this new function to remove apostrophes from function name evaluations for fun(), class_meth, inst_meth, and meth_caller.
* (To make debugging easier, I also added a reference to the ppx_deriving_runtime.  This allowed me to print debug statements of complex objects)

A few existing tests had to be adjusted: function.php.exp and method.php.exp.  Both of those contained some usage of the function-name-eval pattern and their expected files included apostrophes.  This diff changes the expect files to remove apostrophes from the capture groups.

Reviewed By: arxanas

Differential Revision: D13856867

fbshipit-source-id: 757afe4e3e2f57f0378ce6b4e5f69bbf2b61d695