Use `ALoc.t`s more consistently in `Typed_ast_utils`
Summary:
This basically just pushes the conversion later. I added a comment on the function that still uses `Loc.t` explaining why I think this is fine. We can always change it later if it ends up being desirable.

This isn't in support of any specific change, it was just cleanup of `Typed_ast_utils` that I have been meaning to do for a while.

Reviewed By: panagosg7

Differential Revision: D13903776

fbshipit-source-id: d602facf50d57e03947e6d704710233917d54775