Stop the parallelizable workloads loop BEFORE canceling
Summary:
Some users are seeing segfaults when parallelizable jobs run while a recheck is being canceled. In theory the workers should be exiting gracefully and not corrupting the shared memory, but I guess that's not actually happening. The rollback clearly cleans things up, since we don't normally segfault.

So this diff avoids the problem by stopping the parallelizable workload loop a little earlier, before we start canceling workers.

Reviewed By: samwgoldman

Differential Revision: D13942885

fbshipit-source-id: c713c39cf6f8471d43643bd5853f8cd02440a840