[trusted-types] Create a placeholder trust module
Summary: The `trust.ml` and `trust.mli` files have been added, which define the `trust` type and the `bogus_trust` function for creating *ex nihilo* trust information. Currently, `trust` is just a placeholder defined as `unit`, and `bogus_trust` just returns `()`. Future diffs will add more ways of creating `trust` values based on the context of the value's creation (e.g. a literal will be trusted and the `any` type will be untrusted). The intent is for `bogus_trust` to be used when it's not obvious what the correct trust information for a type should be, and one criterion for the trusted types project being ready to ship is when `bogus_trust` is never (or almost never) used.

Reviewed By: dsainati1

Differential Revision: D13743859

fbshipit-source-id: 3b5170334bcc2f7dd3b878b1d09b09e491c142d8