[trusted-types] Add a trust field to all DefTs to prepare for trusted types
Summary:
This diff adds a field to every `DefT` which will contain information about whether the type is "trusted" (does not have `any` flowing into it) or "private" (not flowing into `any`). At the moment this information isn't actually tracked: this diff just adds the location in which it will be tracked in the future, as well as the beginnings of the interface for creating and manipulating trust information.

There are three main changes:
* The definition of `DefT` in `types.ml` is now `DefT of reason * Trust.trust * def_t`.
* Wherever a `DefT` is constructed, the `trust` value supplied either comes from some other `DefT` (if the new `DefT` is semantically the same type as the old one, for example if an existing type's `reason` is being modified but otherwise is remaining the same), or it comes from a call to `bogus_trust` if there isn't an obvious source for the trust value.
* Whenever a `DefT` is pattern matched against, we ignore the `trust` field unless it is getting propagated to another type as above.

This change will affect other people working on Flow's type system by requiring them to deal with the `trust` field in `DefT`s. My suggestion is that for now everyone takes the same approach that I do in the last two bullet points above: propagating existing trust when it's easy and likely sound to do so, and otherwise inventing new trust with `bogus_trust`. The intent is for `bogus_trust` to be used when it's not obvious what the correct trust information for a type should be, and one criterion for the trusted types project being ready to ship is when `bogus_trust` is never (or almost never) used. Of course, right now `bogus_trust` is the only way to create trust values, so it's used everywhere that it's not obvious that trust comes from somewhere else.

Reviewed By: samwgoldman, avikchaudhuri

Differential Revision: D13721310

fbshipit-source-id: eb3fa62c5ad6b4eadc12726e6cfd2f7c4139f62b