Use number of workers, not number of processors, to determine merge bucket size
Summary:
This logic is intended to divide work roughly evenly between the available
workers. Normally the number of workers is the same as the number of
processors, but can be different through configuration.

If the configured number of workers is less than the number of processors, the
existing logic will cause oversmall buckets. If the configured number of
workers is greater than the number of processors, the buckets will be too
large. In both cases, the end result is poor utilization of workers.

Reviewed By: gabelevi

Differential Revision: D13945046

fbshipit-source-id: eabe2403bb689fde783912d9cada971d9be075f0