Update the JSON.stringify
To disallow undefined and functions from being passed in.
JSON.stringify will return undefined if undefined or a function is passed in.

if we allow undefined and functions to be passed in, we would need to change the return value to be `void | string` to be correct. Since that would cause way more trouble for users of JSON.stringify, disallowing the problematic inputs seems like the correct approach here.