make Ast.Expression.Assignment.operator an option
Summary:
the spec does not include `=` in [`AssignmentOperator`](https://tc39.github.io/ecma262/#prod-AssignmentOperator), because pretty much everything about `=` is different than `+=`, `*=`, etc.

we have an `Assign` constructor alongside all the other operators. but it is natural to handle normal `=` assignments differently than assignments with operators, and so making `operator` optional (so `None` represents `=`) makes pattern matching on the operator cleaner and more closely matches the spec as a bonus.

Reviewed By: motiz88

Differential Revision: D13950631

fbshipit-source-id: 8e6c10d6890b1331fe1743301ea1ae385e07e681