Lwtify `Jsonrpc` module to fix possible deadlock
Summary:
The `Jsonrpc` module reads and timestamps messages sent on `stdin` in a non-blocking manner. We didn't have Lwt when we wrote this, so it's a long and complicated `Unix.select`-filled piece of code. (Ideally we would remove it, as Lwt makes it almost trivial, but rearchitecting that is not a priority right now.)

The `Jsonrpc` module functions by creating a daemon process that inherits `stdin` in the background. That process consumes `stdin` and timestamps messages as they come in. In the meantime, we can do blocking work in the client (such as waiting on a server RPC call to finish).

In the client, we sometimes ask the `Jsonrpc` child process if it has any messages to forward to us. We do this by using `Unix.select`; however, we've found that `Unix.select` doesn't play well with `Lwt` because of how signal handling works. There is a sequence of events which caused deadlocks in the serverless IDE code I'm working on:

1. We launch the client.
2. The client launches its `Jsonrpc` child process.
3. We call `Unix.select` on the pipe connecting the two with `0` timeout, indicating that we should return early if nothing is there.
4. Lwt somehow mucks everything up and `Unix.select` is interrupted. For some reason, this causes a retry of the `Unix.select` call, but with an unlimited timeout.
5. We deadlock and cannot do any other Lwt processing while this interminably-blocking `Unix.select` call is underway.

If this has been happening in practice, it's not an issue because:

* It's non-deterministic and doesn't always happen.
* Any action from the user will send a JSONRPC message which will then wake up the client to do more work.
* We weren't really doing much Lwt work in parallel anyways.

We fix this by migrating the `Jsonrpc` module functions which are called in the client to use `Lwt`. This should have been done during the initial Lwt `hh_client` migration, but I missed it at the time.

Note that there are calls to `Unix.select` that we don't need to migrate. Those are launched in the daemon subprocess, so they don't affect or use the Lwt event loop in the main client process.

Reviewed By: gabelevi

Differential Revision: D13776587

fbshipit-source-id: 432e0e68be0d504fe0ddee054af88df7da9bab6f