filter out suppressed lints in workers
Summary:
This adds logic to filter suppressed lints in the workers rather than in the master process at the end of the merge process. The upshot of this is that the `severity_cover` no longer needs to be included in the server environment, thus saving space. Instead, the `errors` set which previously just contained all errors regardless of severity has been split into two sets, `errors` and `warnings` which contain the sorted errors by severity.

We also include the `warnings` map in the saved state, but this is not actually worse for space because it only contains errors that were previously in `local_errors` but are not any more.

Error suppression (which can affect errors and warnings both) still happens at the end, because we still allow error suppressions to affect errors and warnings at any of their locations rather than just the primary one.

Reviewed By: nmote

Differential Revision: D13866166

fbshipit-source-id: a0837d74988a20dd0505914f26882753fb04b26e