Add `require_set` to `File_sig`
Summary: Some callers only care about the list of required modules, and the location of those required modules doesn't matter. Later in the stack, I need to get a collection of the required modules, but it might come from a `File_sig.With_Loc.t` or a `File_sig.With_ALoc.t`, but we don't care about the location at that callsite. So, we need a utility like this.

Reviewed By: jbrown215

Differential Revision: D13962695

fbshipit-source-id: 13969573ee258adb874ad35998db53c312612821