Filter out unenabled lints before constructing the flow_error object
Summary: The call to `error_of_msg` in `add_output` is deceptively expensive, since constructing an error requires a significant amount of additional computation to produce a friendly plaintext string. It is also almost always unnecessary, since we suppress the vast majority of the errors we produce. This changes the lint filtering process to happen without first building the entire error, so we don't have to construct a printable error message for a lint that immediately gets discarded.

Reviewed By: panagosg7

Differential Revision: D14009215

fbshipit-source-id: 4fa18dbaaf39ec2f86054f68d4ff3cd940cf6b53