add lint warning against exported any types
Summary:
This raises a new lint warning (dynamic-exports) when an `any` appears in any position in the exports of a module. Note that this error is added to lib files (in contrast to many others which are not), since a primary motivator of this new lint is changes to lib files that cause coverage to drop in their dependents. However, given that many dynamic types are exported from Flow's standard libdefs, we ignore any instances of this lint raised in the default lib directory.

We also ignore certain specific kinds of implicit `any`s (most relevantly `this` and constructor/prototype types in functions), since these would cause far too much noise since they would trigger whenever a module exports a function. Technically it would still be possible then to export an `any` type through some arcane construction like
```
function f () {};
let x : typeof f.constructor = 0;
module.exports = x;
```
but I can't imagine this is going to be a very common occurrence.

Reviewed By: panagosg7

Differential Revision: D13812096

fbshipit-source-id: 32ce79aeff834809acd640bdfec260ff0571db0a