Include location of JSX children in AST
Summary:
This is needed to avoid reconstructing this location downstream (see the next diffs). This change is made both to normal JSX element children and to JSX fragment children: since they share a lot of code, it was simpler to change both atomically.

Note that nothing currently uses this newly-created location. It will be used in the next diff.

Reviewed By: mroch

Differential Revision: D14016688

fbshipit-source-id: a4c7df83ff09a4bbf694b9926dbafc55e868ebee