Make type parameter to `http$Agent` covariant
Summary:
Followup to D13868636 / https://github.com/facebook/flow/pull/7215

We found a place where someone wanted to use an `http$Agent` from either the `http` module or the `https` module, depending on some runtime flag. Unfortunately, the original change made it so that `http$Agent<net$Socket>` did not have a subtype relationship with `http$Agent<tls$TLSSocket>`.

This change makes it so that `http$Agent<tls$TLSSocket>` is a subtype of `http$Agent<net$Socket>`. I believe that the readonlyness that I added in several places matches the spirit of the API.

Reviewed By: dsainati1

Differential Revision: D14107704

fbshipit-source-id: e2347d6d8c8315a0f263ab6268023c1524475b10