Consolidate merge stream dependents into a single graph structure
Summary:
The merge stream works by incrementally unblocking components when their
dependencies are merged. Before this change, we tracked information about
components in various structures: blocking, recheck_leader_map, component_map,
and dependents. This change consolidates all of that information into a single
graph.

The primary benefit of this change is simplicity. We also need to perform fewer
map/hash table lookups. Finally, the size of the component is part of the node
now, and does not need to be calculated during merge.

Reviewed By: nmote

Differential Revision: D13966593

fbshipit-source-id: 4f1f51eac200c577f347e6f224507a80705270bf