Use queue instead of stack for merge stream
Summary:
The stack behavior causes newly-unblocked jobs to be processed before jobs
already in the stream. This is more like a depth-first traversal of the
dependency graph, when a breadth-first traversal makes more sense.

For example, consider a dependency graph like this (-> means depends on)

B -> A
C -> B
D -> C
X -> A
Y -> X
Z -> Y

Before this change, we would merge A, B, and C before merging X, Y, and Z.
After this change, we'd merge A, then B and X, then C and Y and so on.

Reviewed By: nmote

Differential Revision: D13966509

fbshipit-source-id: e09c0bbef3c9c25924ab63aed20dedc238d25e18