Use unblocked files, not components, to determine bucket sizes
Summary:
In D8758296 I changed the max_bucket_size constant to mean the number of files
in a bucket rather than the number of components. However, I did not correctly
update the logic to calculate the actual bucket size.

The logic error became clear in D13972924 where I refactored the code to use
the name `ready_components` instead of `jobs`. We are comparing files and
components, when we should be comparing files and files.

Reviewed By: nmote

Differential Revision: D13993319

fbshipit-source-id: fdc7d4d7075b6e06fb60618457248cc7dae61479