Add new option --pattern to astCommand to be used by transforms
Summary:
This diff adds a new flag to be used to print the AST structurally. This representation is useful when working with transformations and codemods, to be used in pattern matching.

This new flag is used instead of `--debug` to avoid printing the Locs, as they're just polymorphic in codemods.

Differential Revision: D14113527

fbshipit-source-id: 12f84d2d43de4298a13b3c5973d483f7425fb274