Add support for forwarding leading and trailing nodes in Estree
Summary:
This diff adds support for leading and trailing comments in ESTree nodes. The support is behind a new flag `include_interned_comments` that'll be used to control the rollout to external consumers. It defaults to `false` for now.

It also includes wiring for identifiers as they're the only ones built-in for now. More coming!

Reviewed By: mroch

Differential Revision: D13845267

fbshipit-source-id: 513dc58a3bd1617b84b3bba392fa97dba36a72f3