handle package.json files that are valid JSON but invalid packages
Summary:
if a package.json file is not valid JSON, we raise a parse error. but if it's valid JSON, but not a valid package.json, like if it's an array instead of an object, then we'd hit an `assert_false`.

this diff makes it raise an error on the package.json, and adds an error flag to the package heap so that we can avoid erroring again every time that module is required.

Reviewed By: gabelevi

Differential Revision: D13960310

fbshipit-source-id: 4e4a00ca2856666be158813760276eada8a36468