[Refactor] rename `error` to `printable_error`
Summary:
In order to distinguish formatted, printable `error`s and bare-bones `error_message`s from the new intermediate type that will be introduced, this diff renames the original `error` type to `printable_error` to denote that it should not be used until we are ready to print it. It also renames the type's associated modules and destructor functions to reflect its new name.

Most of the places where we use the renamed `printable_error` type or an associated module will be refactored to use the intermediate representation in upcoming diffs.

Reviewed By: gabelevi

Differential Revision: D14030588

fbshipit-source-id: ef73623d36cf40bcc70d10f7536db51d1f1da561