store `Flow_error.t`s in the context and the server instead of `Errors.printable_error`s
Summary:
Prior to this change, the pipeline for creating and outputting errors in Flow looked roughly like this:

Identify error -> Create printable error -> Save error in context -> Filter suppressed lints-> Save errors to server -> Concretize errors -> Filter suppressed errors -> Print errors

This is inefficient for a number of reasons, but the most important one is that making human readable errors is a lot of work that is almost never worth it, considering that we suppress almost every single error we produce. The formatting is particularly slow, and saving raw strings in the context and server is an enormous waste of memory.

This diff changes this pipeline to:

Identify error -> Create error -> Save error in context -> Filter suppressed lints -> Save errors to server -> Concretize errors -> Make errors printable -> Filter suppressed errors -> Print errors

This avoids storing the raw string error message in the context or the server, and in the case of lints skips the formatting step for lints that would be filtered out anyways.

Note that we still filter suppressed errors after they are made printable; this is because errors can be suppressed at any of their locations, and the logic to determine what counts as a location for each error is dependent on which locations appear in the raw text error message. If we change the suppression behavior to only suppress at the primary location, we can rearrange the end of the pipeline to filter suppressed errors before making them printable, presumably saving even more time.

Reviewed By: gabelevi

Differential Revision: D14115493

fbshipit-source-id: 7a60360454e6d6239a6d04e934d4b80e2267e418