[PR] Update the JSON.stringify
Summary:
To disallow undefined and functions from being passed in.
JSON.stringify will return undefined if undefined or a function is passed in.

if we allow undefined and functions to be passed in, we would need to change the return value to be `void | string` to be correct. Since that would cause way more trouble for users of JSON.stringify, disallowing the problematic inputs seems like the correct approach here.

<!--
  If this is a change to library defintions, please include links to relevant documentation.
  If this is a documentation change, please prefix the title with [DOCS].

  If this is neither, ensure you opened a discussion issue and link it in the PR description.
-->
Pull Request resolved: https://github.com/facebook/flow/pull/7447

Reviewed By: nmote

Differential Revision: D13975675

Pulled By: nmn

fbshipit-source-id: a749f3adf234794b188f596f7f362726d1964e45