make version_regex tolerate prereleases
Summary:
this function takes flow's version number and turns it into a regex that matches all lower versions. we use this for the `<VERSION>` component of suppression comments.

this diff makes it tolerant of prerelease identifiers in flow's version, like `0.94.0-beta.1`. it will only look at the major, minor and patch.

in fact, since this is only for matching against flow's version number, which is always a full major, minor, patch, we can remove the tests for fewer components and only `sscanf` for `%u.%u.%u`.

Reviewed By: gabelevi

Differential Revision: D14185544

fbshipit-source-id: 668a85f88320a3ed701fd975497f3454211f0d1b