[semver] handle prerelease and build identifiers
Summary:
adds support for prerelease identifiers, so that we can publish versions like `0.94.0-beta.1` and match ranges like `^0.94.0-alpha` in the .flowconfig `[version]`.

the format and precedence is based on the [Semantic Versioning 2.0.0](https://semver.org/) spec, and the way it interacts with ranges (the `^` or `>`) follows what `node-semver` does (see [here](https://github.com/npm/node-semver#prerelease-tags)).

Reviewed By: gabelevi

Differential Revision: D14183825

fbshipit-source-id: 0b2fd4f1e536e933a18919a38b552269e133101a